Function to calculate the vaccine coverage per dose

This function returns the vaccination coverage of a dose along the cohort study. The coverage can be calculated grouped by year, day and month. This most by specified in the parameter unit. If there are not registers for some dates, the function assigns 0, instead of NA, to be able to calculate the cumulative coverage.

Usage

coh_coverage(
  data,
  vacc_date_col,
  unit = c("day", "month", "year"),
  date_interval = NULL
)

Arguments

data

dataset with cohort information (see example)

vacc_date_col

name of the column(s) that contains the vaccine date to calculate the coverage

unit

aggregation unit, must be either "year" or "month" or "day"

date_interval

if NULL, the function calculates the coverage interval based on the min() and max() of the vacc_date_col. It is also possible to pass a custom date interval to truncate or expand the date interval (see example)

Value

data.frame with vaccine number of doses per date, cumulative count of doses and vaccine coverage

Examples

data("cohortdata")
start_cohort <- as.Date("2044-01-01")
end_cohort <- as.Date("2044-12-31")
date_interval <- c(start_cohort, end_cohort)
coh_coverage(
  data = cohortdata,
  vacc_date_col = "vaccine_date_1",
  unit = "month",
  date_interval = date_interval
)
#>          date doses cum_doses coverage
#> 1  2044-01-01     0         0  0.00000
#> 2  2044-02-01   512       512  0.00512
#> 3  2044-03-01  4129      4641  0.04641
#> 4  2044-04-01  3051      7692  0.07692
#> 5  2044-05-01  7041     14733  0.14733
#> 6  2044-06-01 10920     25653  0.25653
#> 7  2044-07-01  4808     30461  0.30461
#> 8  2044-08-01  3125     33586  0.33586
#> 9  2044-09-01  2134     35720  0.35720
#> 10 2044-10-01  1059     36779  0.36779
#> 11 2044-11-01  2282     39061  0.39061
#> 12 2044-12-01    66     39127  0.39127