Function to Estimate the Vaccine Effectiveness Using Hazard Ratio

The function relies on the implementation of the Cox model for proportional hazards from the package {survival} The proportional hazards hypothesis is tested using the Schoenfeld test, and the resultant p-value is provided in the results. Log-log plots are also calculated using the Kaplan-Meier survival estimator to provide a visual test for the proportional hazards hypothesis. The function returns a list with the method called, the estimation of VE (CI95%), the p-value of the Schoenfeld test, and the log-log plot.

Usage

coh_eff_hr(
  data_set,
  outcome_status_col,
  time_to_event_col,
  vacc_status_col,
  vaccinated_status,
  unvaccinated_status,
  start_cohort,
  end_cohort
)

Arguments

data_set

data.frame with cohort information (see example).

outcome_status_col

Name of the column containing status of the event (must be a binary column). Default is outcome_status.

time_to_event_col

Name of the column containing the time-to-event. Default is time_to_event.

vacc_status_col

Name of the column containing the vaccination. Default is vaccine_status.

vaccinated_status

Status assigned to the vaccinated population. Default is v.

unvaccinated_status

Status assigned to the unvaccinated population. Default is u.

start_cohort

Start date of the study.

end_cohort

End date of the study.

Value

Method call from {survival}, VE (CI95%), p-value of Schoenfeld test, and log-log plot for proportional hazards