Censor couple after matching

This function censors a couple whether the case or the control have a censoring date. It imputes the censoring date to the whole couple using the matching id provided in subclass. This column comes with the output of match_cohort.

Usage

get_censoring_date_match(data, outcome_date_col, censoring_date_col)

Arguments

data

dataset with cohort information (see example)

outcome_date_col

name of the column that contains the outcome dates

censoring_date_col

name of the column that contains the censoring date. NULL by default

Examples

# load package example data for cohort studies
data("cohortdata")

# assign vaccination status
cohortdata$vaccine_status <- set_status(
  data = cohortdata,
  col_names = c("vaccine_date_1", "vaccine_date_2"),
  status = c("v", "u")
)

# match cohort
matched_cohort <- match_cohort(data = cohortdata,
  status_vacc_col = "vaccine_status",
  nearest = c(age = 1),
  exact = "sex"
)

# add column with censoring date for match
matched_cohort$censoring_date_match <-  get_censoring_date_match(
  data = matched_cohort,
  outcome_date_col = "death_date",
  censoring_date_col = "death_other_causes"
)

# view data with added column
head(matched_cohort)
#>         id sex age death_date death_other_causes vaccine_date_1 vaccine_date_2
#> 1 afade1b2   F  37       <NA>               <NA>           <NA>           <NA>
#> 3 04edf85a   M  50       <NA>               <NA>           <NA>           <NA>
#> 5 c5a83f56   M  66       <NA>               <NA>           <NA>           <NA>
#> 6 7f675ec3   M  29       <NA>               <NA>     2044-04-09     2044-04-30
#> 7 5bec597e   F  22       <NA>               <NA>           <NA>           <NA>
#> 9 47507fa7   M  15       <NA>               <NA>     2044-08-22     2044-11-03
#>   vaccine_1 vaccine_2 vaccine_status prop_score subclass censoring_date_match
#> 1      <NA>      <NA>              u  0.4109251    29591                 <NA>
#> 3      <NA>      <NA>              u  0.4448482    29675                 <NA>
#> 5      <NA>      <NA>              u  0.5845737    30430                 <NA>
#> 6    BRAND1    BRAND1              v  0.2767662    21686                 <NA>
#> 7      <NA>      <NA>              u  0.2915153       84                 <NA>
#> 9    BRAND1    BRAND1              v  0.1895003    34827                 <NA>