Function to construct the inmunization date

This function returns a column with the immunization date that corresponds to an analyzed outcome. The function searches for vaccine dates that satisfy the condition: vacc_date_col <= outcome_date_col - delay_time - immunization_delay. This condition allows to discriminate the vaccine dates in terms of characteristic times in days (delay_time) associated to an outcome, from the onset of symptoms or from any reference event, and the characteristic time in days before the patient is considered immune (immunization_delay). Both parameters can be set to zero by the user without affecting the results. When take_first = FALSE, the function uses the vaccine date that is closest to the outcome to calculate the immunization. On the other hand, when take_first = TRUE the functions uses the first vaccine date.

Usage

get_immunization_date(
  data,
  outcome_date_col,
  censoring_date_col = NULL,
  outcome_delay,
  immunization_delay,
  vacc_date_col,
  end_cohort,
  take_first = FALSE
)

Arguments

data

dataset with cohort information (see example)

outcome_date_col

name of the column that contains the outcome dates

censoring_date_col

name of the column that contains the censoring date. NULL by default

outcome_delay

characteristic time in days of the outcome from the reference event

immunization_delay

characteristic time in days before the patient is considered immune

vacc_date_col

name of the column(s) that contains the vaccine dates

end_cohort

end date of the study

take_first

TRUE: takes the minimum vaccine date . FALSE: takes the closest to end_cohort

Value

immunization date

Details

If a register does not present an outcome, the function searches for vaccine dates that satisfy the condition: vacc_date_col <= end_cohort - delay_time - immunization_delay. The immunization date is constructed using the closest vaccine date to the end of the study, when take_first = FALSE; or the first vaccination date found, when take_first = TRUE. Notice that the function works for one or several vaccines. In case of several vaccines, the parameter outcome_date_col must be passed as a vector (see example).

Examples

# load package example data
data("cohortdata")
# define end date of the study as type date
end_cohort <- as.Date("2044-12-31")
# get immunization dates
cohortdata$immunization <- get_immunization_date(
  data = cohortdata,
  outcome_date_col = "death_date",
  outcome_delay = 0,
  immunization_delay = 14,
  vacc_date_col = c("vaccine_date_1", "vaccine_date_2"),
  end_cohort = end_cohort,
  take_first = FALSE
)

# view head of data
head(cohortdata)
#>         id sex age death_date death_other_causes vaccine_date_1 vaccine_date_2
#> 1 afade1b2   F  37       <NA>               <NA>           <NA>           <NA>
#> 2 556c8c76   M  19       <NA>               <NA>           <NA>           <NA>
#> 3 04edf85a   M  50       <NA>               <NA>           <NA>           <NA>
#> 4 7e51a18e   F   8       <NA>               <NA>           <NA>           <NA>
#> 5 c5a83f56   M  66       <NA>               <NA>           <NA>           <NA>
#> 6 7f675ec3   M  29       <NA>               <NA>     2044-04-09     2044-04-30
#>   vaccine_1 vaccine_2 immunization
#> 1      <NA>      <NA>         <NA>
#> 2      <NA>      <NA>         <NA>
#> 3      <NA>      <NA>         <NA>
#> 4      <NA>      <NA>         <NA>
#> 5      <NA>      <NA>         <NA>
#> 6    BRAND1    BRAND1   2044-05-14