Function to construct vaccine biologic associated to the immunization date

This function returns the names of the vaccines associated to the immunization date. To avoid mistakes, it is necessary to set the same value of immunization_delay that was used in the previous functions. The arguments vacc_date_col and vacc_name_col must be passed in the same order, i.e. every name of column date must correspond to a name of vaccine column (see example)

Usage

get_immunization_vaccine(
  data,
  immunization_date_col,
  vacc_date_col,
  vacc_name_col,
  immunization_delay
)

Arguments

data

dataset with cohort information (see example)

immunization_date_col

name of the column that contains the immunization date.

vacc_date_col

name of the column(s) that contains the vaccine date

vacc_name_col

name of the column(s) that contains the vaccine names

immunization_delay

characteristic time in days before the patient is considered immune

Value

dose: a column with the names of the columns that are associated to the doses of each register

vaccine name

Examples

# load example data
data("cohortdata")
cohortdata$immunization <- get_immunization_date(
  data = cohortdata,
  outcome_date_col = "death_date",
  outcome_delay = 0,
  immunization_delay = 14,
  vacc_date_col = c("vaccine_date_1", "vaccine_date_2"),
  end_cohort = as.Date("2044-12-31"),
  take_first = FALSE
)
cohortdata$immunization_vaccine <- get_immunization_vaccine(
  data = cohortdata,
  immunization_date_col = "immunization",
  vacc_date_col = c("vaccine_date_1", "vaccine_date_2"),
  vacc_name_col = c("vaccine_1", "vaccine_2"),
  immunization_delay = 14
)

# view data
head(cohortdata)
#>         id sex age death_date death_other_causes vaccine_date_1 vaccine_date_2
#> 1 afade1b2   F  37       <NA>               <NA>           <NA>           <NA>
#> 2 556c8c76   M  19       <NA>               <NA>           <NA>           <NA>
#> 3 04edf85a   M  50       <NA>               <NA>           <NA>           <NA>
#> 4 7e51a18e   F   8       <NA>               <NA>           <NA>           <NA>
#> 5 c5a83f56   M  66       <NA>               <NA>           <NA>           <NA>
#> 6 7f675ec3   M  29       <NA>               <NA>     2044-04-09     2044-04-30
#>   vaccine_1 vaccine_2 immunization immunization_vaccine
#> 1      <NA>      <NA>         <NA>                 <NA>
#> 2      <NA>      <NA>         <NA>                 <NA>
#> 3      <NA>      <NA>         <NA>                 <NA>
#> 4      <NA>      <NA>         <NA>                 <NA>
#> 5      <NA>      <NA>         <NA>                 <NA>
#> 6    BRAND1    BRAND1   2044-05-14               BRAND1