Function to construct the time-to-event

This function returns a column with the time-to-event in days occurred until a reference outcome. The starting point to count the time-to-event can be the immunization date, supposing that the vaccinate population enters to the study, when they are vaccinated (start_from_immunization=TRUE). Or the beginning of the study, if all the cohort is known at this point (start_from_immunization=FALSE). In this last case, it is not necessary to pass the argument immunization_date_col

Usage

get_time_to_event(
  data,
  outcome_date_col,
  censoring_date_col = NULL,
  start_cohort,
  end_cohort,
  start_from_immunization = FALSE,
  immunization_date_col
)

Arguments

data

dataset with cohort information (see example)

outcome_date_col

name of the column that contains the outcome dates

censoring_date_col

name of the column that contains the censoring date. NULL by default

start_cohort

start date of the study

end_cohort

end date of the study

start_from_immunization

TRUE: starts counting time-to-event from immunization date if available FALSE: starts counting time-to-event from the start date of the cohort study

immunization_date_col

name of the column that contains the immunization date. Required if start_from_immunization = TRUE

Value

time-to-event

Examples

# load package example data
data("cohortdata")
# define start and end dates of the study as type date
start_cohort <- as.Date("2044-01-01")
end_cohort <- as.Date("2044-12-31")
cohortdata$immunization <- get_immunization_date(
  data = cohortdata,
  outcome_date_col = "death_date",
  outcome_delay = 0,
  immunization_delay = 14,
  vacc_date_col = c("vaccine_date_1", "vaccine_date_2"),
  end_cohort = end_cohort,
  take_first = FALSE
)
cohortdata$time_to_death <- get_time_to_event(
  data = cohortdata,
  outcome_date_col = "death_date",
  start_cohort = start_cohort,
  end_cohort = end_cohort,
  start_from_immunization = TRUE,
  immunization_date_col = "immunization"
)

# view data
head(cohortdata)
#>         id sex age death_date death_other_causes vaccine_date_1 vaccine_date_2
#> 1 afade1b2   F  37       <NA>               <NA>           <NA>           <NA>
#> 2 556c8c76   M  19       <NA>               <NA>           <NA>           <NA>
#> 3 04edf85a   M  50       <NA>               <NA>           <NA>           <NA>
#> 4 7e51a18e   F   8       <NA>               <NA>           <NA>           <NA>
#> 5 c5a83f56   M  66       <NA>               <NA>           <NA>           <NA>
#> 6 7f675ec3   M  29       <NA>               <NA>     2044-04-09     2044-04-30
#>   vaccine_1 vaccine_2 immunization time_to_death
#> 1      <NA>      <NA>         <NA>           365
#> 2      <NA>      <NA>         <NA>           365
#> 3      <NA>      <NA>         <NA>           365
#> 4      <NA>      <NA>         <NA>           365
#> 5      <NA>      <NA>         <NA>           365
#> 6    BRAND1    BRAND1   2044-05-14           231