Static match of cohort

This function builds couples of vaccinated - unvaccinated individuals with similar characteristics. The function relies on the propensity score matching algorithm implemented in MatchIt package. By default the function uses method = "nearest", ratio = 1, distance = "glm" to match the data. Exact and near characteristics are accepted for the matching criteria. These are passed in the parameters exact and nearest, respectively. Parameters nearest and caliper must be provided together. In this case, the calipers must be passed as a named vector containing each of the variables provided in nearest (e.g. nearest = c("characteristic1", "characteristic2"), caliper = c(characteristic1 = n1, characteristic2 = n2), where n1 and n2 are the calipers). caliper is ignored (set to NULL) when nearest is not provided.

Usage

match_cohort(data, status_vacc_col, exact = NULL, nearest = NULL)

Arguments

data

dataset with cohort information (see example)

status_vacc_col

name of the column containing the information of the vaccination status.

exact

name(s) of column(s) for exact matching. Default to NULL.

nearest

named vector with name(s) of column(s) for nearest matching and caliper(s) for each variable. e.g. nearest = c("characteristic1" = n1, "characteristic2" = n2), where n1 and n2 are the calipers. Default to NULL.

Value

data frame with matched population. Two columns are added to the structure provided in data: prop_score (propensity score of the match), subclass (id of matched couple)

Examples

# load package example data for cohort studies
data("cohortdata")

# assign vaccination status
cohortdata$vaccine_status <- set_status(
  data = cohortdata,
  col_names = c("vaccine_date_1", "vaccine_date_2"),
  status = c("v", "u")
)

# match cohort
matched_cohort <- match_cohort(data = cohortdata,
  status_vacc_col = "vaccine_status",
  nearest = c(age = 1),
  exact = "sex"
)

# view matched data
head(matched_cohort)
#>         id sex age death_date death_other_causes vaccine_date_1 vaccine_date_2
#> 1 afade1b2   F  37       <NA>               <NA>           <NA>           <NA>
#> 3 04edf85a   M  50       <NA>               <NA>           <NA>           <NA>
#> 5 c5a83f56   M  66       <NA>               <NA>           <NA>           <NA>
#> 6 7f675ec3   M  29       <NA>               <NA>     2044-04-09     2044-04-30
#> 7 5bec597e   F  22       <NA>               <NA>           <NA>           <NA>
#> 9 47507fa7   M  15       <NA>               <NA>     2044-08-22     2044-11-03
#>   vaccine_1 vaccine_2 vaccine_status prop_score subclass
#> 1      <NA>      <NA>              u  0.4109251    29591
#> 3      <NA>      <NA>              u  0.4448482    29675
#> 5      <NA>      <NA>              u  0.5845737    30430
#> 6    BRAND1    BRAND1              v  0.2767662    21686
#> 7      <NA>      <NA>              u  0.2915153       84
#> 9    BRAND1    BRAND1              v  0.1895003    34827