Calculate Vaccine Coverage

This function returns the vaccination coverage of a dose along the cohort study. The coverage can be calculated grouped by year, day, and month. This must be specified in the parameter unit. If there are no records for some dates, the function assigns 0 instead of NA to calculate the cumulative coverage.

Usage

coh_coverage(
  data_set,
  vacc_date_col,
  unit = c("day", "month", "year"),
  date_interval = NULL
)

Arguments

data_set

data.frame with cohort information (see example).

vacc_date_col

Name of the column(s) that contain the vaccine date to calculate the coverage.

unit

Aggregation unit, must be either "year", "month", or "day".

date_interval

If NULL, the function calculates the coverage interval based on the min() and max() of the vacc_date_col. It is also possible to pass a custom date interval to truncate or expand the date interval (see example).

Value

data.frame with the number of vaccine doses per date, cumulative count of doses, and vaccine coverage.

Examples

# Load example data
data("cohortdata")

# Define date intervals for coverage
start_cohort <- as.Date("2044-01-01")
end_cohort <- as.Date("2044-12-31")
date_interval <- c(start_cohort, end_cohort)

# Calculate coverage
coh_coverage(
  data_set = cohortdata,
  vacc_date_col = "vaccine_date_1",
  unit = "month",
  date_interval = date_interval
)
#>          date doses cum_doses coverage
#> 1  2044-01-01     0         0  0.00000
#> 2  2044-02-01   512       512  0.00512
#> 3  2044-03-01  4129      4641  0.04641
#> 4  2044-04-01  3051      7692  0.07692
#> 5  2044-05-01  7041     14733  0.14733
#> 6  2044-06-01 10920     25653  0.25653
#> 7  2044-07-01  4808     30461  0.30461
#> 8  2044-08-01  3125     33586  0.33586
#> 9  2044-09-01  2134     35720  0.35720
#> 10 2044-10-01  1059     36779  0.36779
#> 11 2044-11-01  2282     39061  0.39061
#> 12 2044-12-01    66     39127  0.39127