Function to test the proportional hazards hypothesis of the model.

The function relies on the implementation of the Schoenfeld test for testing the proportional hazards hypothesis on survival package. It returns the result of the test and it can by plotted using plot(). If the test is rejected, it is recomended to try other statistical strategies, e.g. stratifying the dataset by confounders or including time-dependent variables.

Usage

coh_test_noconf(
  data,
  outcome_status_col,
  time_to_event_col,
  status_vacc_col,
  p_thr = 0.05
)

Arguments

data

dataset with cohort information (see example)

outcome_status_col

name of the column containing status of the event (most be a binary column)

time_to_event_col

name of the column containing the time-to-event

status_vacc_col

name of the column containing the vaccination status

p_thr

p-value to test Proportional Hazards Hypothesis

Value

Schoenfeld test: Survival element

Examples

# load example package data
data("cohortdata")

# add useful columns
cohortdata$immunization <- get_immunization_date(
  data = cohortdata,
  outcome_date_col = "death_date",
  outcome_delay = 0,
  immunization_delay = 14,
  vacc_date_col = c("vaccine_date_1", "vaccine_date_2"),
  end_cohort = as.Date("2044-12-31"),
  take_first = FALSE
)

cohortdata$vaccine_status <- set_status(
  data = cohortdata,
  col_names = c("immunization"),
  status = c("v", "u")
)

cohortdata$death_status <- set_status(
  data = cohortdata,
  col_names = c("death_date")
)

cohortdata$time_to_death <- get_time_to_event(
  data = cohortdata,
  outcome_date_col = "death_date",
  start_cohort = as.Date("2044-01-01"),
  end_cohort = as.Date("2044-12-31"),
  start_from_immunization = FALSE
)

# perform the test
test <- coh_test_noconf(
  cohortdata,
  "death_status",
  "time_to_death",
  "vaccine_status"
)

# View test object
test
#>                         chisq df       p
#> data[[status_vacc_col]]    11  1 0.00091
#> GLOBAL                     11  1 0.00091