Function to construct the time-to-event

This function returns a column with the time-to-event in days occurred until a reference outcome. The starting point to count the time-to-event can be the immunization date, supposing that the vaccinate population enters to the study, when they are vaccinated (start_from_immunization=TRUE). Or the beginning of the study, if all the cohort is known at this point (start_from_immunization=FALSE). In this last case, it is not necessary to pass the argument immunization_date_col

Usage

get_time_to_event(
  data,
  outcome_date_col,
  start_cohort,
  end_cohort,
  start_from_immunization = FALSE,
  immunization_date_col
)

Arguments

data

dataset with cohort information (see example)

outcome_date_col

name of the column that contains the outcome dates

start_cohort

start date of the study

end_cohort

end date of the study

start_from_immunization

TRUE: starts counting time-to-event from immunization date if available FALSE: starts counting time-to-event from the start date of the cohort study

immunization_date_col

name of the column that contains the immunization date. Required if start_from_immunization = TRUE

Value

time-to-event

Examples

# load package example data
data("cohortdata")
# define start and end dates of the study as type date
start_cohort <- as.Date("2044-01-01")
end_cohort <- as.Date("2044-12-31")
cohortdata$immunization <- get_immunization_date(
  data = cohortdata,
  outcome_date_col = "death_date",
  outcome_delay = 0,
  immunization_delay = 14,
  vacc_date_col = c("vaccine_date_1", "vaccine_date_2"),
  end_cohort = end_cohort,
  take_first = FALSE
)
cohortdata$time_to_death <- get_time_to_event(
  data = cohortdata,
  outcome_date_col = "death_date",
  start_cohort = start_cohort,
  end_cohort = end_cohort,
  start_from_immunization = TRUE,
  immunization_date_col = "immunization"
)

# view data
head(cohortdata)
#>   sex age subsidy death_date vaccine_date_1 vaccine_date_2 vaccine_1 vaccine_2
#> 1   F   6       0       <NA>           <NA>           <NA>      <NA>      <NA>
#> 2   M  79       0       <NA>     2044-03-31     2044-05-07    BRAND2    BRAND2
#> 3   F  34       0       <NA>     2044-07-26     2044-09-03    BRAND2    BRAND2
#> 4   M  26       0       <NA>           <NA>           <NA>      <NA>      <NA>
#> 5   F  66       0       <NA>     2044-05-20     2044-06-17    BRAND2    BRAND2
#> 6   M  10       1       <NA>           <NA>           <NA>      <NA>      <NA>
#>   immunization time_to_death
#> 1         <NA>           365
#> 2   2044-05-21           224
#> 3   2044-09-17           105
#> 4         <NA>           365
#> 5   2044-07-01           183
#> 6         <NA>           365