Assign outcome status or vaccination status

This function generates a binary status column using the set of columns, passed through the variable col_names. This column must contain the information of the outcomes or the vaccine dates. It generates a logical condition using the operators & or |, and evaluates if the registers in the columns contain or not information. If the logical operator is &, the function returns a success, only if all the columns contain information. On the other hand, if the logical operator is |, it is enough to find one column with information indicating success. It is recommended to use this method when working with several outcomes or several vaccine doses. By default, it returns a binary column where 0 means no outcome or no vaccine and 1 means the opposite. However, it can also receive personalized options, e.g. c("v", "u") for vaccinated and unvaccinated.

Usage

set_status(data, col_names, operator = c("&", "|"), status = c(1, 0))

Arguments

data

Dataset, as a data.frame with at least one column from which to generate the status specified in status.

col_names

Name or names of the column(s) as a string or a character vector containing the information from which the status is calculated.

operator

A single logical operator to evaluate the condition.

status

A two-element vector specifying the values to be assigned that indicate whether the individual is vaccinated or not e.g. c("v","u"). The first element of the vector must be the status when the condition is satisfied, i.e., vaccinated, while the second element is the value indicating that the individual is not vaccinated.

Value

status

Examples

# load package example data for cohort studies
data("cohortdata")

# assign vaccination status
cohortdata$vaccine_status <- set_status(
  data = cohortdata,
  col_names = c("vaccine_date_1", "vaccine_date_2"),
  status = c("v", "u")
)

# view data with added column
head(cohortdata)
#>         id sex age death_date death_other_causes vaccine_date_1 vaccine_date_2
#> 1 afade1b2   F  37       <NA>               <NA>           <NA>           <NA>
#> 2 556c8c76   M  19       <NA>               <NA>           <NA>           <NA>
#> 3 04edf85a   M  50       <NA>               <NA>           <NA>           <NA>
#> 4 7e51a18e   F   8       <NA>               <NA>           <NA>           <NA>
#> 5 c5a83f56   M  66       <NA>               <NA>           <NA>           <NA>
#> 6 7f675ec3   M  29       <NA>               <NA>     2044-04-09     2044-04-30
#>   vaccine_1 vaccine_2 vaccine_status
#> 1      <NA>      <NA>              u
#> 2      <NA>      <NA>              u
#> 3      <NA>      <NA>              u
#> 4      <NA>      <NA>              u
#> 5      <NA>      <NA>              u
#> 6    BRAND1    BRAND1              v